fix(transport): fix issues related path, map and concurrency
Previous version doesn't care about the filename is nil case and
the number of matched file is 0 case. This will expect to raise the
serious problem. I fix of it.

In addition, I found that the current mapping scheme suffers serious space
overhead. Therefore, I add the compaction routine in the transport
package.

Event worse is that the concurrency issue was detected at naive-bayesian
model. This issue makes us to prepare for unforeseen consequences. I
hate this kinds of issue. Nevertheless, mutex will generate the some
overhead but I thought it to trade-off.